fix possible :ProjectTab <dir> error
When you already have 2 project tabs open, for different projects, and
attempt to issue a :ProjectTab <dir> command, where <dir> is relative to
the cwd of the project in the first tab, that dir was not expanded,
resulting in an error during the Tcd stage of the process. This is
because the new tab has the working directory of the previous tab in the
list.